+++ include/minishell.h
@@ -6,7 +6,7 @@
 /*   By: porellan <porellan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/18 12:35:21 by miniore           #+#    #+#             */
-/*   Updated: 2025/08/13 20:05:56 by porellan         ###   ########.fr       */
+/*   Updated: 2025/08/13 21:02:45 by porellan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -92,6 +92,7 @@ int		ft_is_quotes(char c);
 int		ft_is_dquotes(char c);
 
 void	free_array(char **array);
+void free_command_list(tok_lst *commands_lst, size_t size);
 char    *ft_strjoin_free(char *s1, char *s2);
 void    ft_put_pererr(t_backpack *backpack, char *err, int n);
 void    ft_put_syserr(t_backpack *backpack, char *err);
