+++ src/build_ins/export.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   export.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: frlorenz <frlorenz@student.42.fr>          +#+  +:+       +#+        */
+/*   By: porellan <porellan@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/27 19:04:42 by porellan          #+#    #+#             */
-/*   Updated: 2025/08/05 20:17:46 by frlorenz         ###   ########.fr       */
+/*   Updated: 2025/08/13 21:03:15 by porellan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,11 +43,31 @@ static int  ft_valid_char(char *var)
     return(EXIT_SUCCESS);
 }
 
+static void ft_save_var(t_backpack *backpack, t_list *actual, char *var)
+{
+    char    *content;
+
+    content = ft_var_content(actual->content);  
+    if (!content)
+        content = "";
+    if (!search_node(&backpack->env, var))
+        env_add_last(&backpack->env, new_node(var, content));
+    else
+        modify_node(search_node(&backpack->env, var), NULL, content);
+    free(var);
+}
+
+static void ft_check_error(t_backpack *backpack, char *var)
+{
+    if((!ft_isalpha(var[0]) && var[0] != '_') || ft_valid_char(var))
+        free(var);
+    ft_put_pererr(backpack, "Minichelita: export: not a valid identifier.\n", 258);
+}
+
 void    ft_export(t_backpack *backpack)
 {
     t_list  *actual;
     char    *var;
-    char    *content;
 
     backpack->n = 0;
     actual = backpack->commands_lst[backpack->n].arguments;
@@ -63,12 +83,14 @@ void    ft_export(t_backpack *backpack)
         var = name_var(actual->content);
         if(!var || (!ft_isalpha(var[0]) && var[0] != '_') || ft_valid_char(var))
         {
-            if((!ft_isalpha(var[0]) && var[0] != '_') || ft_valid_char(var))
-                free(var);
-            ft_put_pererr(backpack, "Minichelita: export: not a valid identifier.\n", 258);
+            ft_check_error(backpack, var);
             actual = actual->next;
             continue ;
         }
+        ft_save_var(backpack, actual, var);
+        actual = actual->next;
+    }
+}
         // if(ft_valid_char(var))
         // {
         //     ft_put_pererr(backpack, "Minichelita: export: not a valid identifier.\n", 258);
@@ -76,15 +98,3 @@ void    ft_export(t_backpack *backpack)
         //     actual = actual->next;
         //     continue ;
         // }
-        content = ft_var_content(actual->content);  
-        if (!content)
-            content = "";
-        if (!search_node(&backpack->env, var))
-            env_add_last(&backpack->env, new_node(var, content));
-        else
-            modify_node(search_node(&backpack->env, var), NULL, content);
-        free(var);
-        actual = actual->next;
-    }
-    
-}
